refactor: change video package location
