+++ src/main/java/com/scg/stop/user/domain/User.java
@@ -5,14 +5,13 @@ import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.*;
 import static lombok.AccessLevel.*;
 
-import com.scg.stop.auth.domain.request.RegisterRequest;
 import com.scg.stop.domain.project.domain.Comment;
 import com.scg.stop.domain.project.domain.FavoriteProject;
 import com.scg.stop.domain.project.domain.Inquiry;
 import com.scg.stop.domain.project.domain.Likes;
 import com.scg.stop.domain.proposal.domain.Proposal;
-import com.scg.stop.domain.video.domain.FavoriteVideo;
-import com.scg.stop.domain.video.domain.UserQuiz;
+import com.scg.stop.video.domain.FavoriteVideo;
+import com.scg.stop.video.domain.UserQuiz;
 import com.scg.stop.global.domain.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
