+++ src/main/java/com/scg/stop/video/domain/JobInterview.java
@@ -1,22 +1,25 @@
-package com.scg.stop.domain.video.domain;
+package com.scg.stop.video.domain;
 
+import static jakarta.persistence.EnumType.STRING;
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PRIVATE;
 import static lombok.AccessLevel.PROTECTED;
 
-import com.scg.stop.domain.video.dto.request.TalkRequest;
 import com.scg.stop.global.domain.BaseTimeEntity;
 import jakarta.persistence.*;
 
 import java.util.List;
+
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
+@AllArgsConstructor(access = PRIVATE)
 @NoArgsConstructor(access = PROTECTED)
-public class Talk extends BaseTimeEntity {
+public class JobInterview extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
@@ -37,57 +40,53 @@ public class Talk extends BaseTimeEntity {
     @Column(nullable = false)
     private String talkerName;
 
-    @OneToOne(fetch = LAZY, mappedBy = "talk", cascade = CascadeType.REMOVE, orphanRemoval = true)
-    private Quiz quiz;
+    @Column(nullable = false)
+    @Enumerated(value = STRING)
+    private JobInterviewCategory category;
 
-    @OneToMany(fetch = LAZY, mappedBy = "talk")
+    @OneToMany(fetch = LAZY, mappedBy = "jobInterview", cascade = CascadeType.REMOVE, orphanRemoval = true)
     private List<FavoriteVideo> favoriteVideos;
 
-    public Talk(String title, String youtubeId, Integer year, String talkerBelonging, String talkerName) {
+    private JobInterview(String title, String youtubeId, Integer year, String talkerBelonging, String talkerName, JobInterviewCategory category) {
         this.title = title;
         this.youtubeId = youtubeId;
         this.year = year;
         this.talkerBelonging = talkerBelonging;
         this.talkerName = talkerName;
+        this.category = category;
     }
-    public static Talk from(
+
+    public static JobInterview from(
             String title,
             String youtubeId,
             Integer year,
             String talkerBelonging,
-            String talkerName
+            String talkerName,
+            JobInterviewCategory category
     ) {
-        return new Talk(
+        return new JobInterview(
                 title,
                 youtubeId,
                 year,
                 talkerBelonging,
-                talkerName
+                talkerName,
+                category
         );
     }
 
-    public void updateTalk(
+    public void updateJobInterview(
             String title,
             String youtubeId,
             Integer year,
             String talkerBelonging,
-            String talkerName
+            String talkerName,
+            JobInterviewCategory category
     ) {
-        this.title= title;
+        this.title = title;
         this.youtubeId = youtubeId;
         this.year = year;
         this.talkerBelonging = talkerBelonging;
         this.talkerName = talkerName;
+        this.category = category;
     }
-
-    public void setQuiz(Quiz quiz) {
-        if(this.quiz != null) {
-            this.quiz.setTalk(null);
-        }
-        this.quiz = quiz;
-        if(quiz != null) {
-            quiz.setTalk(this);
-        }
-    }
-
 }
