+++ src/main/java/com/scg/stop/video/domain/UserQuiz.java
@@ -1,4 +1,4 @@
-package com.scg.stop.domain.video.domain;
+package com.scg.stop.video.domain;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
@@ -6,6 +6,7 @@ import static lombok.AccessLevel.PROTECTED;
 
 import com.scg.stop.user.domain.User;
 import com.scg.stop.global.domain.BaseTimeEntity;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
@@ -17,21 +18,23 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class FavoriteVideo extends BaseTimeEntity {
+public class UserQuiz extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "talk_id")
-    private Talk talk;
+    @Column(nullable = false)
+    private Integer tryCount = 1;
+
+    @Column(nullable = false, columnDefinition = "TINYINT(1)")
+    private boolean isSuccess = false;
 
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "job_interview_id")
-    private JobInterview jobInterview;
-
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "user_ide")
+    @JoinColumn(name = "user_id")
     private User user;
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "quiz_id")
+    private Quiz quiz;
 }
