+++ src/main/java/com/scg/stop/video/dto/request/QuizInfoRequest.java
@@ -1,10 +1,9 @@
-package com.scg.stop.domain.video.dto.request;
+package com.scg.stop.video.dto.request;
 
-import com.scg.stop.domain.video.domain.QuizInfo;
+import com.scg.stop.video.domain.QuizInfo;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
