+++ src/main/java/com/scg/stop/video/repository/QuizRepository.java
@@ -1,6 +1,6 @@
-package com.scg.stop.domain.video.repository;
+package com.scg.stop.video.repository;
 
-import com.scg.stop.domain.video.domain.Quiz;
+import com.scg.stop.video.domain.Quiz;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuizRepository extends JpaRepository<Quiz, Long> {
