+++ src/test/java/com/scg/stop/domain/video/controller/JobInterviewControllerTest.java
@@ -2,12 +2,11 @@ package com.scg.stop.domain.video.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.scg.stop.configuration.AbstractControllerTest;
-import com.scg.stop.domain.video.domain.JobInterview;
-import com.scg.stop.domain.video.domain.JobInterviewCategory;
-import com.scg.stop.domain.video.dto.request.JobInterviewRequest;
-import com.scg.stop.domain.video.dto.response.JobInterviewResponse;
-import com.scg.stop.domain.video.repository.JobInterviewRepository;
-import com.scg.stop.domain.video.service.JobInterviewService;
+import com.scg.stop.video.controller.JobInterviewController;
+import com.scg.stop.video.domain.JobInterviewCategory;
+import com.scg.stop.video.dto.request.JobInterviewRequest;
+import com.scg.stop.video.dto.response.JobInterviewResponse;
+import com.scg.stop.video.service.JobInterviewService;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
