+++ src/test/java/com/scg/stop/domain/video/controller/TalkControllerTest.java
@@ -2,12 +2,13 @@ package com.scg.stop.domain.video.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.scg.stop.configuration.AbstractControllerTest;
-import com.scg.stop.domain.video.dto.request.QuizInfoRequest;
-import com.scg.stop.domain.video.dto.request.QuizRequest;
-import com.scg.stop.domain.video.dto.request.TalkRequest;
-import com.scg.stop.domain.video.dto.response.QuizResponse;
-import com.scg.stop.domain.video.dto.response.TalkResponse;
-import com.scg.stop.domain.video.service.TalkService;
+import com.scg.stop.video.controller.TalkController;
+import com.scg.stop.video.dto.request.QuizInfoRequest;
+import com.scg.stop.video.dto.request.QuizRequest;
+import com.scg.stop.video.dto.request.TalkRequest;
+import com.scg.stop.video.dto.response.QuizResponse;
+import com.scg.stop.video.dto.response.TalkResponse;
+import com.scg.stop.video.service.TalkService;
 import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
